AI_Legal_Explainer/requirements_phase4.txt
# Phase 4: Production & Deployment Requirements

# Security & Compliance
cryptography==41.0.7
django-ratelimit==4.1.0
django-security==0.16.0
django-csp==3.7
django-honeypot==1.0.2
django-axes==6.3.0

# Testing & Quality Assurance
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-html==4.1.1
selenium==4.15.2
locust==2.17.0
coverage==7.3.2

# Performance & Monitoring
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-silk==5.1.0
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Documentation & API
drf-spectacular==0.27.0
django-rest-swagger==2.2.0
markdown==3.5.1
mkdocs==1.5.3

# Production & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
celery==5.3.4
redis==5.0.1

# Security Testing
bandit==1.7.5
safety==2.3.5
pip-audit==2.6.1

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Additional Utilities
python-decouple==3.8
django-environ==0.11.2
psutil==5.9.6
